requirements.txt
autopep8==1.5.7
Django==2.2.19
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dotenv==0.18.0
pytz==2021.1
reportlab==3.5.68
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
